Crossbow is firing!

+++ fortress/src/lib/map/map_config.rs
@@ -1,9 +1,9 @@
 #[derive(Clone, Deserialize)]
 pub struct Platform {
-    pub top_left_x: i32,
-    pub top_left_y: i32,
-    pub width: i32,
-    pub height: i32
+    pub top_left_x: f32,
+    pub top_left_y: f32,
+    pub width: f32,
+    pub height: f32
 }
 
 #[derive(Clone, Deserialize)]
